config: add option ssh-authorized-keys
Introduce a new config option to set the path to the authorized keys
file that will be used to write the ssh public key of the user for
remote access. The config option also supports the usage of common
variables $HOME, $USER and $UID which will be expanded for each user.

Closes-bug: 1983437

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>